fix: delete discount

+++ src/services/admin/discount.service.ts
@@ -2,6 +2,8 @@ import {Discount} from "../../models/discount";
 import { RedisCache } from "../../config/redis-cache";
 import { ObjectId } from "mongodb";
 import { Product } from "../../models/product";
+import { ProductDetail } from "../../models/product-detail";
+import { CartDetail } from "../../models/cart-detail";
 export class DiscountService {
 
     static async getAllDiscount(){
@@ -110,13 +112,36 @@ export class DiscountService {
         try {
             
             const discount = await Discount.findOne({_id:discountId});
+            let now = new Date();
+            now.setHours(now.getHours()+7);
+            if(discount.startDate.getTime()<now.getTime()&&discount.endDate.getTime()>now.getTime()){
+                return {status: 400, message:"Bạn không thể xóa giảm giá này"}
+            }
             // const products = await Product.find({discount: new ObjectId(`${discountId}`)})
             await Discount.deleteOne({_id: new ObjectId(`${discountId}`)});
             const key = `getAllDiscount()`;
+            const products = await Product.find({'discount': new ObjectId(`${discountId}`)});
+            for (let index = 0; index < products.length; index++) {
+                let productDetails = await ProductDetail.aggregate([{$match:{product:new ObjectId(`${products[index]._id}`)}},{$match:{status:"ACTIVE"}},{$project:{_id:1}}])
+                let productDetailsConvert = productDetails.map(function(id) {
+                    return id;
+                });
+                await CartDetail.updateMany({productDetail:{$in:productDetailsConvert}},{$set:{priceDiscount:products[index].price}});
+            }
+            await Product.updateMany({'discount': new ObjectId(`${discountId}`)},{$set:{'discount':new ObjectId('62599849f8f6be052f0a901d')}})
+            delKeyRedisWhenDeleteDiscount();
             await RedisCache.delCache(key);
             return {status: 200, message:"delete discount success !"};
         } catch (error) {
             return{status:500,message: "Something went wrong !", error: error};
         }
     }
+}
+async function delKeyRedisWhenDeleteDiscount() {
+    const keysCart = await RedisCache.getKeys(`CartService*`);
+    await RedisCache.delKeys(keysCart);
+    const keysProduct = await RedisCache.getKeys(`ProductService*`);
+    await RedisCache.delKeys(keysProduct);
+    const keysFavorite = await RedisCache.getKeys(`FavoriteService*`);
+    await RedisCache.delKeys(keysFavorite);
 }
